Move tests/fakes.py into pyro/distributions/testing
