+++ pyro/distributions/testing/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module contains distributions that are useful for testing new Pyro
+inference algorithms.
+"""
