+++ pyro/distributions/testing/fakes.py
