+++ tests/integration_tests/test_tracegraph_elbo.py
@@ -13,10 +13,10 @@ from torch.nn import Parameter
 import pyro
 import pyro.distributions as dist
 import pyro.optim as optim
+from pyro.distributions.testing import fakes
 from pyro.distributions.transformed_distribution import TransformedDistribution
 from pyro.infer import SVI
 from pyro.util import ng_ones, ng_zeros
-from tests import fakes
 from tests.common import assert_equal
 from tests.distributions.test_transformed_distribution import AffineExp
